support-hints-agent-system/requirements.txt
fastapi
langchain-core
langchain-openai
langgraph
pandas
pydantic
python-dotenv
qdrant-client
rank-bm25
requests
sentence-transformers
spacy
torch
tqdm
uvicorn
openpyxl

# Spacy model
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc